Refactor and relocate dashboard breadcrumb header component
